add mysportfeed module
